+++ CHANGELOG.md
@@ -5,4 +5,4 @@
 
 ### Features
 
-* setting up project dependencies ([#1](https://github.com/Stackclash/obsidian-ttrpg-audio-controller/issues/1)) ([897be79](https://github.com/Stackclash/obsidian-ttrpg-audio-controller/commit/897be796363381d0c49d25f05c616d5de1eb9b41))
+* setting up project dependencies ([#1](https://github.com/Stackclash/obsidian-ttrpg-audio-manager/issues/1)) ([897be79](https://github.com/Stackclash/obsidian-ttrpg-audio-manager/commit/897be796363381d0c49d25f05c616d5de1eb9b41))